LayoutTests/fast/dom/mutation-events-enablement-expected.txt

This tests disabling mutation events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mutationEvents is ["DOMNodeInserted", "DOMNodeInsertedIntoDocument", "DOMSubtreeModified", "DOMNodeRemoved", "DOMNodeRemovedFromDocument", "DOMSubtreeModified"]
PASS mutationEvents is ["DOMCharacterDataModified", "DOMSubtreeModified"]
PASS successfullyParsed is true

TEST COMPLETE
world
